Change tests for elliptic redistancing
