+++ proteus/tests/elliptic_redist/RDLS/clean.sh
