+++ proteus/tests/elliptic_redist/RDLS/ncls_n.py
@@ -1,54 +1,64 @@
 from proteus import *
 from proteus.default_n import *
-from rdls_p import *
+from ncls_p import *
 from vortex2D import *
+nd = 2
 
-timeIntegration = NoIntegration
-stepController = Newton_controller
+# About time integration 
+timeIntegration = BackwardEuler_cfl
+stepController = Min_dt_controller
 
-# About the nonlinear solver
+# About nonlinear solver
 multilevelNonlinearSolver  = Newton
 levelNonlinearSolver = Newton
 
+nl_atol_res = atolLevelSet
+l_atol_res = atolLevelSet
 tolFac = 0.0
-nl_atol_res = atolRedistance
-linTolFac = 0.0
 
-maxNonlinearIts = 100000
+maxNonlinearIts = 25
 maxLineSearches = 0
-useEisenstatWalker = True
 fullNewtonFlag = True
 
 if useHex:
     hex=True
-    if pDegree_ls==1:
+    if pDegree_ls == 1:
         femSpaces = {0:C0_AffineLinearOnCubeWithNodalBasis}
-    elif pDegree_ls==2:
+    elif pDegree_ls == 2:
         femSpaces = {0:C0_AffineLagrangeOnCubeWithNodalBasis}
+    else:
+        print "pDegree_ls = %s not recognized " % pDegree_ls
     elementQuadrature = CubeGaussQuadrature(nd,vortex_quad_order)
     elementBoundaryQuadrature = CubeGaussQuadrature(nd-1,vortex_quad_order)
 else:
-    if pDegree_ls==1:
+    if pDegree_ls == 1:
         femSpaces = {0:C0_AffineLinearOnSimplexWithNodalBasis}
-    elif pDegree_ls==2:
+    elif pDegree_ls == 2:
         femSpaces = {0:C0_AffineQuadraticOnSimplexWithNodalBasis}
+    else:
+        print "pDegree_ls = %s not recognized " % pDegree_ls
     elementQuadrature = SimplexGaussQuadrature(nd,vortex_quad_order)
     elementBoundaryQuadrature = SimplexGaussQuadrature(nd-1,vortex_quad_order)
 
-subgridError = HamiltonJacobi_ASGS_opt(coefficients,nd,stabFlag='2',lag=False)
-shockCapturing = RDLS.ShockCapturing(coefficients,nd,shockCapturingFactor=shockCapturingFactor_rd,lag=lag_shockCapturing_rd)
+subgridError = HamiltonJacobi_ASGS_opt(coefficients,nd,lag=True)
+massLumping = False
+numericalFluxType = None
 
+shockCapturing = NCLS.ShockCapturing(coefficients,
+                                     nd,
+                                     shockCapturingFactor=shockCapturingFactor_ls,
+                                     lag=lag_shockCapturing_ls)
 numericalFluxType = DoNothing
 
+nonlinearSolverConvergenceTest = 'rits'
+levelNonlinearSolverConvergenceTest = 'rits'
 nonlinearSmoother = None
-levelNonlinearSolverConvergenceTest='r'
-nonlinearSolverConvergenceTest='r'
 
 matrix = SparseMatrix
 if parallel:
     multilevelLinearSolver = KSP_petsc4py#PETSc
     levelLinearSolver = KSP_petsc4py#PETSc
-    linear_solver_options_prefix = 'rdls_'
+    linear_solver_options_prefix = 'ncls_'
     linearSolverConvergenceTest = 'r-true'
 else:
     multilevelLinearSolver = LU
