+++ proteus/tests/elliptic_redist/RDLS/test_redist.py
@@ -34,9 +34,9 @@ class TestEllipticRedistancing():
 
     def test_ELLIPTIC_REDISTANCING_0(self):
         # Set parameters for test #
-        vortex2D.ct.ELLIPTIC_REDISTANCING = False
+        vortex2D.ct.ELLIPTIC_REDISTANCING = 0
         reload(rdls_p)
-        reload(vortex2D_so)
+        reload(vortex2D_so)        
         pnList = [(ncls_p,ncls_n),
                   (rdls_p,rdls_n)]
         self.so = vortex2D_so
@@ -59,13 +59,12 @@ class TestEllipticRedistancing():
                                                opts)
         ns.calculateSolution('rdls')
         actual = tables.open_file('vortex_c0p1_level_1_ELLIPTIC_REDIST_0.h5','r')
-        assert np.isclose(np.amax(actual.root.u_t2),0.137239911295,atol=1e-10)
+        assert np.isclose(np.amax(actual.root.u_t1),0.1338208537581557,atol=1e-10)
         actual.close()
 
     def test_ELLIPTIC_REDISTANCING_1(self):
         # Set parameters for test #
-        vortex2D.ct.ELLIPTIC_REDISTANCING = True
-        vortex2D.ct.ELLIPTIC_REDISTANCING_TYPE = 1
+        vortex2D.ct.ELLIPTIC_REDISTANCING = 1
         reload(rdls_p)
         reload(vortex2D_so)
         pnList = [(ncls_p,ncls_n),
@@ -90,13 +89,12 @@ class TestEllipticRedistancing():
                                                opts)
         ns.calculateSolution('rdls')
         actual = tables.open_file('vortex_c0p1_level_1_ELLIPTIC_REDIST_1.h5','r')
-        assert np.isclose(np.amax(actual.root.u_t2),0.144279865323,atol=1e-10)
+        assert np.isclose(np.amax(actual.root.u_t1),0.1304760274072587,atol=1e-10)
         actual.close()
 
     def test_ELLIPTIC_REDISTANCING_2(self):
         # Set parameters for test #
-        vortex2D.ct.ELLIPTIC_REDISTANCING = True
-        vortex2D.ct.ELLIPTIC_REDISTANCING_TYPE = 2
+        vortex2D.ct.ELLIPTIC_REDISTANCING = 2
         reload(default_p)
         reload(default_so)
         reload(rdls_p)
@@ -123,13 +121,13 @@ class TestEllipticRedistancing():
                                                opts)
         ns.calculateSolution('rdls')
         actual = tables.open_file('vortex_c0p1_level_1_ELLIPTIC_REDIST_2.h5','r')
-        assert np.isclose(np.amax(actual.root.u_t2),0.112863525457,atol=1e-10)
+        print np.amax(actual.root.u_t1)
+        assert np.isclose(np.amax(actual.root.u_t1),0.10626477076036571,atol=1e-10)
         actual.close()
 
     def test_ELLIPTIC_REDISTANCING_3(self):
         # Set parameters for test #
-        vortex2D.ct.ELLIPTIC_REDISTANCING = True
-        vortex2D.ct.ELLIPTIC_REDISTANCING_TYPE = 3
+        vortex2D.ct.ELLIPTIC_REDISTANCING = 3
         reload(rdls_p)
         reload(vortex2D_so)
         pnList = [(ncls_p,ncls_n),
@@ -154,5 +152,5 @@ class TestEllipticRedistancing():
                                                opts)
         ns.calculateSolution('rdls')
         actual = tables.open_file('vortex_c0p1_level_1_ELLIPTIC_REDIST_3.h5','r')
-        assert np.isclose(np.amax(actual.root.u_t2),0.112863525462,atol=1e-10)
+        assert np.isclose(np.amax(actual.root.u_t1),0.10598184134747586,atol=1e-10)
         actual.close()
