+++ proteus/tests/elliptic_redist/RDLS/vortex2D.py
@@ -1,17 +1,16 @@
 from proteus import Domain
 from proteus import Context
-from proteus.mprans import NCLS3P
+from proteus.mprans import NCLS
 import numpy as np
 
 ct=Context.Options([
     # General parameters #
-    ("T",0.1,"Final time"),
+    ("T",0.01,"Final time"),
     ("nDTout",1,"Number of time steps to archive"),
     ("refinement",1,"Level of refinement"),
     ("unstructured",False,"Use unstructured mesh. Set to false for periodic BCs"),
     # parameters for elliptic re-distancing    
-    ("ELLIPTIC_REDISTANCING",False, "Type of elliptic re-distancing"),
-    ("alpha",100.0,"penalization parameter")
+    ("ELLIPTIC_REDISTANCING",0, "Type of elliptic re-distancing"),
 ],mutable=True)
 
 # ELLIPTIC_REDISTANCING #
@@ -25,7 +24,7 @@ runCFL=0.33
 nd=2
 
 # General parameters 
-parallel = False
+parallel = True
 linearSmoother = None
 checkMass=False
 
@@ -83,10 +82,9 @@ fmmFlag=0
 
 soname="vortex_c0p"+`pDegree_ls`+"_level_"+`ct.refinement`
 
-class MyCoefficients(NCLS3P.Coefficients):
+class MyCoefficients(NCLS.Coefficients):
     def attachModels(self,modelList):
         self.model = modelList[0]
         self.q_v = np.zeros(self.model.q[('dH',0,0)].shape,'d')
         self.ebqe_v = np.zeros(self.model.ebqe[('dH',0,0)].shape,'d')
-        self.rdModel = None
-        self.rdModel_ebqe = np.copy(self.model.ebqe[('u',0)])
+        self.rdModel = self.model
