+++ proteus/tests/elliptic_redist/RDLS3P/clean.sh
