+++ proteus/tests/elliptic_redist/RDLS3P/ncls_n.py
@@ -65,6 +65,3 @@ else:
     levelLinearSolver = LU
 
 conservativeFlux = {}
-
-
-
