+++ proteus/tests/elliptic_redist/RDLS3P/test_redist.py
@@ -12,8 +12,8 @@ import numpy as np
 import tables
 import pytest
 from . import (vortex2D, vortex2D_so,
-               ncls3P_p, ncls3P_n,
-               rdls3P_p, rdls3P_n)
+               ncls_p, ncls_n,
+               rdls_p, rdls_n)
 
 class TestEllipticRedistancing():
 
@@ -35,10 +35,10 @@ class TestEllipticRedistancing():
     def test_ELLIPTIC_REDISTANCING_0(self):
         # Set parameters for test #
         vortex2D.ct.ELLIPTIC_REDISTANCING = 0
-        reload(rdls3P_p)
-        reload(vortex2D_so)
-        pnList = [(ncls3P_p,ncls3P_n),
-                  (rdls3P_p,rdls3P_n)]
+        reload(rdls_p)
+        reload(vortex2D_so)        
+        pnList = [(ncls_p,ncls_n),
+                  (rdls_p,rdls_n)]
         self.so = vortex2D_so
         pList=[]
         nList=[]
@@ -57,18 +57,18 @@ class TestEllipticRedistancing():
                                                nList,
                                                sList,
                                                opts)
-        ns.calculateSolution('rdls3P')
+        ns.calculateSolution('rdls')
         actual = tables.open_file('vortex_c0p1_level_1_ELLIPTIC_REDIST_0.h5','r')
-        assert np.isclose(np.amax(actual.root.u_t2),0.137547581396,atol=1e-10)
+        assert np.isclose(np.amax(actual.root.u_t1),0.1338208537581557,atol=1e-10)
         actual.close()
 
     def test_ELLIPTIC_REDISTANCING_1(self):
         # Set parameters for test #
         vortex2D.ct.ELLIPTIC_REDISTANCING = 1
-        reload(rdls3P_p)
+        reload(rdls_p)
         reload(vortex2D_so)
-        pnList = [(ncls3P_p,ncls3P_n),
-                  (rdls3P_p,rdls3P_n)]
+        pnList = [(ncls_p,ncls_n),
+                  (rdls_p,rdls_n)]
         self.so = vortex2D_so
         pList=[]
         nList=[]
@@ -87,18 +87,20 @@ class TestEllipticRedistancing():
                                                nList,
                                                sList,
                                                opts)
-        ns.calculateSolution('rdls3P')
-        actual = tables.open_file('vortex_c0p1_level_1_ELLIPTIC_REDIST_1.h5','r')        
-        assert np.isclose(np.amax(actual.root.u_t2),0.144686270399,atol=1e-10)
+        ns.calculateSolution('rdls')
+        actual = tables.open_file('vortex_c0p1_level_1_ELLIPTIC_REDIST_1.h5','r')
+        assert np.isclose(np.amax(actual.root.u_t1),0.1304760274072587,atol=1e-10)
         actual.close()
 
     def test_ELLIPTIC_REDISTANCING_2(self):
         # Set parameters for test #
         vortex2D.ct.ELLIPTIC_REDISTANCING = 2
-        reload(rdls3P_p)
+        reload(default_p)
+        reload(default_so)
+        reload(rdls_p)
         reload(vortex2D_so)
-        pnList = [(ncls3P_p,ncls3P_n),
-                  (rdls3P_p,rdls3P_n)]
+        pnList = [(ncls_p,ncls_n),
+                  (rdls_p,rdls_n)]
         self.so = vortex2D_so
         pList=[]
         nList=[]
@@ -117,18 +119,19 @@ class TestEllipticRedistancing():
                                                nList,
                                                sList,
                                                opts)
-        ns.calculateSolution('rdls3P')
-        actual = tables.open_file('vortex_c0p1_level_1_ELLIPTIC_REDIST_2.h5','r')        
-        assert np.isclose(np.amax(actual.root.u_t2),0.113047815938,atol=1e-10)
+        ns.calculateSolution('rdls')
+        actual = tables.open_file('vortex_c0p1_level_1_ELLIPTIC_REDIST_2.h5','r')
+        print np.amax(actual.root.u_t1)
+        assert np.isclose(np.amax(actual.root.u_t1),0.10626477076036571,atol=1e-10)
         actual.close()
 
     def test_ELLIPTIC_REDISTANCING_3(self):
         # Set parameters for test #
         vortex2D.ct.ELLIPTIC_REDISTANCING = 3
-        reload(rdls3P_p)
+        reload(rdls_p)
         reload(vortex2D_so)
-        pnList = [(ncls3P_p,ncls3P_n),
-                  (rdls3P_p,rdls3P_n)]
+        pnList = [(ncls_p,ncls_n),
+                  (rdls_p,rdls_n)]
         self.so = vortex2D_so
         pList=[]
         nList=[]
@@ -147,8 +150,7 @@ class TestEllipticRedistancing():
                                                nList,
                                                sList,
                                                opts)
-        ns.calculateSolution('rdls3P')
+        ns.calculateSolution('rdls')
         actual = tables.open_file('vortex_c0p1_level_1_ELLIPTIC_REDIST_3.h5','r')
-        print np.amax(actual.root.u_t2)
-        assert np.isclose(np.amax(actual.root.u_t2),0.113047815938,atol=1e-10)
+        assert np.isclose(np.amax(actual.root.u_t1),0.10598184134747586,atol=1e-10)
         actual.close()
