+++ setup.py
@@ -557,7 +557,8 @@ def setup_given_extensions(extensions):
                       'proteus.tests.BernsteinPolynomials',
                       'proteus.tests.BernsteinPolynomials.poisson_eqn',
                       'proteus.tests.elliptic_redist',
-                      'proteus.tests.elliptic_redist_3P',
+                      'proteus.tests.elliptic_redist.RDLS',
+                      'proteus.tests.elliptic_redist.RDLS3P',
                       'proteus.tests.surface_tension',
                       'proteus.tests.surface_tension.rising_bubble_rans3p',
                       'proteus.tests.CLSVOF',
